requirements.txt
numpy
torch
ultralytics
matplotlib
opencv-python
tqdm
pillow
pandas